Safari-611.1.10/LayoutTests/media/video-controls-in-media-document-expected.txt

Test that controls don't increase the size of the container (i.e. are rendered overlapping with the video canvas).



EXPECTED (controls.offsetTop + controls.offsetHeight <= '240') OK
